feat(237-04): extract dataset search orchestration
- move search_datasets, FTS ranking, search-only filters, and sort resolution into service_datasets.py

- keep SearchFilters and search_datasets available through the public search facade
